+++ .github/workflows/cicd.yaml
@@ -24,17 +24,17 @@ jobs:
       uses: actions/checkout@v2
 
     - name: Build docker image
-      run: docker build -t lidar_deep_im .
+      run: docker build -t myria3d .
 
     - name: Check code neatness (linter)
-      run: docker run lidar_deep_im flake8
+      run: docker run myria3d flake8
 
     - name: Fast unit tests first
       run: >
         docker run
-        -v /var/data/CICD_github_assets/nebulia_B2V0.5/inputs/:/lidar/tests/files/large/
+        -v /var/data/CICD_github_assets/nebulia_B2V0.5/inputs/:/myria3d/tests/files/large/
         --ipc=host
-        lidar_deep_im
+        myria3d
         python -m
         pytest -rA -v -m "not slow"
         --ignore=actions-runner
@@ -45,24 +45,24 @@ jobs:
         -v /var/data/cicd/CICD_github_assets/nebulia_B2V0.5/inputs/:/inputs/
         -v /var/data/cicd/CICD_github_assets/nebulia_B2V0.5/outputs/:/outputs/
         --ipc=host
-        lidar_deep_im
+        myria3d
         python -m 
-        lidar_multiclass.predict
+        myria3d.predict
         --config-path /inputs/
         --config-name predict_config_V2.0.0.yaml
         predict.src_las=/inputs/792000_6272000_subset_buildings.las
         predict.output_dir=/outputs/
         predict.ckpt_path=/inputs/RandLaNet_Buildings_B2V0.5_epoch_033.ckpt
         datamodule.batch_size=25
-        hydra.run.dir=/lidar
+        hydra.run.dir=/myria3d
 
     # Always run with --ipc=host to enable sufficient shared memory when predicting on large data
     - name: Slow unit tests last, using large files stored in self-hosted action-runner.
       run: >
         docker run
-        -v /var/data/cicd/CICD_github_assets/nebulia_B2V0.5/inputs/:/lidar/tests/data/large/
+        -v /var/data/cicd/CICD_github_assets/nebulia_B2V0.5/inputs/:/myria3d/tests/data/large/
         --ipc=host
-        lidar_deep_im
+        myria3d
         python -m
         pytest -rA -v -m "slow" 
         --ignore=actions-runner
@@ -80,16 +80,16 @@ jobs:
 
     - name: Tag the docker image with branch name
       if: github.event_name == 'push'
-      run: docker tag lidar_deep_im:latest lidar_deep_im:${BRANCH_NAME}
+      run: docker tag myria3d:latest myria3d:${BRANCH_NAME}
 
     - name: Run the new, tagged image at least once so that is it not prunned by mistake when using docker system prune
       if: github.event_name == 'push'
-      run: docker run lidar_deep_im:${BRANCH_NAME} bash
+      run: docker run myria3d:${BRANCH_NAME} bash
 
     # This needs writing rights to the mounted path
-    - name: Save the docker image as lidar_deep_im_${BRANCH_NAME}.tar
+    - name: Save the docker image as myria3d_${BRANCH_NAME}.tar
       if: github.event_name == 'push'
-      run: docker save lidar_deep_im:${BRANCH_NAME} -o /var/data/cicd/CICD_github_assets/CICD_docker_images/lidar_deep_im_${BRANCH_NAME}.tar
+      run: docker save myria3d:${BRANCH_NAME} -o /var/data/cicd/CICD_github_assets/CICD_docker_images/myria3d_${BRANCH_NAME}.tar
 
     - name: Clean dangling docker images
       if: always()  # always do it, even if something failed
