+++ configs/callbacks/default.yaml
@@ -1,11 +1,11 @@
 # Override with "none" to deactivate. 
 
 log_code:
-  _target_: lidar_multiclass.callbacks.comet_callbacks.LogCode
-  code_dir: ${work_dir}/lidar_multiclass
+  _target_: myria3d.callbacks.comet_callbacks.LogCode
+  code_dir: ${work_dir}/myria3d
 
 log_logs_dir:
-  _target_: lidar_multiclass.callbacks.comet_callbacks.LogLogsPath
+  _target_: myria3d.callbacks.comet_callbacks.LogLogsPath
 
 lr_monitor:
   _target_:  pytorch_lightning.callbacks.LearningRateMonitor
@@ -15,10 +15,10 @@ lr_monitor:
 # This logs IoU at validation and test time
 # Predictions are aggregated and saved at test time in a way coherent with prediction logic.
 log_iou_by_class:
-  _target_: lidar_multiclass.callbacks.logging_callbacks.LogIoUByClass
+  _target_: myria3d.callbacks.logging_callbacks.LogIoUByClass
   classification_dict: ${datamodule.dataset_description.classification_dict}
   interpolator:  # only used at test time
-    _target_: lidar_multiclass.models.interpolation.Interpolator
+    _target_: myria3d.models.interpolation.Interpolator
     interpolation_k: ${predict.interpolation_k}
     classification_dict: ${datamodule.dataset_description.classification_dict}
     probas_to_save: ${predict.probas_to_save}  # replace by a list of string of class names to select specific probas to save
