+++ configs/callbacks/finetuning.yaml
@@ -3,7 +3,7 @@ defaults:
 
 # Override 
 finetune:
-  _target_: lidar_multiclass.callbacks.finetuning_callbacks.FinetuningFreezeUnfreeze
+  _target_: myria3d.callbacks.finetuning_callbacks.FinetuningFreezeUnfreeze
   d_in: ${model.d_in}
   num_classes: ${model.num_classes}
   unfreeze_fc_end_epoch: ???  # full end fc is unfrozen later than the last MLP.
