+++ configs/datamodule/datamodule.yaml
@@ -1,4 +1,4 @@
-_target_: lidar_multiclass.data.datamodule.DataModule
+_target_: myria3d.data.datamodule.DataModule
 
 # Directory with test, train, val prepared subdirs
 prepared_data_dir: ${oc.env:PREPARED_DATA_DIR}
