+++ configs/datamodule/dataset_description/20220204_BuildingValidation_and_Ground.yaml
@@ -8,8 +8,8 @@ classification_dict: {1: "unclassified", 2: "ground", 6: "building"}
 d_in: 12
 num_classes: 3
 
-# loading functions for data object prepared via lidar_multiclass/datamodule/data.py
+# loading functions for data object prepared via myria3d/datamodule/data.py
 load_las_func:
   _target_: functools.partial
   _args_:
-    - "${get_method:lidar_multiclass.data.loading.FrenchLidarDataLogic.load_las}"
+    - "${get_method:myria3d.data.loading.FrenchLidarDataLogic.load_las}"
