+++ configs/datamodule/dataset_description/20220504_proto23dalles.yaml
@@ -8,9 +8,9 @@ classification_dict: {1: "unclassified", 2: "ground", 5: vegetation, 6: "buildin
 d_in: 12
 num_classes: 10
 
-# loading functions for data object prepared via lidar_multiclass/datamodule/data.py
+# loading functions for data object prepared via myria3d/datamodule/data.py
 # Nota: In this function, artefacts as well as synthetic points (65, 66) are filtered out.
 load_las_func:
   _target_: functools.partial
   _args_:
-    - "${get_method:lidar_multiclass.data.loading.FrenchLidarDataLogic.load_las}"
+    - "${get_method:myria3d.data.loading.FrenchLidarDataLogic.load_las}"
