+++ configs/datamodule/subsampler/fps.yaml
@@ -1,2 +1,2 @@
-_target_: lidar_multiclass.data.transforms.FPSSampler
+_target_: myria3d.data.transforms.FPSSampler
 subsample_size: 12500
