+++ configs/datamodule/subsampler/grid.yaml
@@ -1,3 +1,3 @@
-_target_: lidar_multiclass.data.transforms.CustomGridSampler
+_target_: myria3d.data.transforms.CustomGridSampler
 subsample_size: 12500
 voxel_size: 0.25
