+++ docs/source/apidoc/default_config.yml
@@ -30,12 +30,12 @@ datamodule:
     load_las_func:
       _target_: functools.partial
       _args_:
-      - ${get_method:lidar_multiclass.data.loading.FrenchLidarDataLogic.load_las}
+      - ${get_method:myria3d.data.loading.FrenchLidarDataLogic.load_las}
   subsampler:
-    _target_: lidar_multiclass.data.transforms.CustomGridSampler
+    _target_: myria3d.data.transforms.CustomGridSampler
     subsample_size: 12500
     voxel_size: 0.25
-  _target_: lidar_multiclass.data.datamodule.DataModule
+  _target_: myria3d.data.datamodule.DataModule
   prepared_data_dir: ${oc.env:PREPARED_DATA_DIR}
   test_data_dir: ${datamodule.prepared_data_dir}/test/
   batch_size: 16
@@ -46,19 +46,19 @@ datamodule:
   subsample_size: 12500
 callbacks:
   log_code:
-    _target_: lidar_multiclass.callbacks.comet_callbacks.LogCode
-    code_dir: ${work_dir}/lidar_multiclass
+    _target_: myria3d.callbacks.comet_callbacks.LogCode
+    code_dir: ${work_dir}/myria3d
   log_logs_dir:
-    _target_: lidar_multiclass.callbacks.comet_callbacks.LogLogsPath
+    _target_: myria3d.callbacks.comet_callbacks.LogLogsPath
   lr_monitor:
     _target_: pytorch_lightning.callbacks.LearningRateMonitor
     logging_interval: step
     log_momentum: true
   log_iou_by_class:
-    _target_: lidar_multiclass.callbacks.logging_callbacks.LogIoUByClass
+    _target_: myria3d.callbacks.logging_callbacks.LogIoUByClass
     classification_dict: ${datamodule.dataset_description.classification_dict}
     interpolator:
-      _target_: lidar_multiclass.models.interpolation.Interpolator
+      _target_: myria3d.models.interpolation.Interpolator
       interpolation_k: ${predict.interpolation_k}
       classification_dict: ${datamodule.dataset_description.classification_dict}
       probas_to_save: ${predict.probas_to_save}
@@ -94,7 +94,7 @@ model:
     patience: 5
     cooldown: 0
     verbose: true
-  _target_: lidar_multiclass.models.model.Model
+  _target_: myria3d.models.model.Model
   d_in: ${datamodule.dataset_description.d_in}
   num_classes: ${datamodule.dataset_description.num_classes}
   ckpt_path: null
