+++ docs/source/conf.py
@@ -98,6 +98,6 @@ intersphinx_mapping = {
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-modindex_common_prefix = ["lidar_multiclass."]
+modindex_common_prefix = ["myria3d."]
 
 mock_directives = ["testcode"]
