+++ myria3d/predict.py
@@ -6,8 +6,8 @@ from omegaconf import DictConfig
 from pytorch_lightning import LightningDataModule, LightningModule
 from tqdm import tqdm
 
-from lidar_multiclass.utils import utils
-from lidar_multiclass.models.interpolation import Interpolator
+from myria3d.utils import utils
+from myria3d.models.interpolation import Interpolator
 
 
 log = utils.get_logger(__name__)
@@ -76,8 +76,8 @@ def main(config: DictConfig):
     """
     # Imports should be nested inside @hydra.main to optimize tab completion
     # Read more here: https://github.com/facebookresearch/hydra/issues/934
-    from lidar_multiclass.utils import utils
-    from lidar_multiclass.predict import predict
+    from myria3d.utils import utils
+    from myria3d.predict import predict
 
     # You can safely get rid of this line if you don't want those
     utils.extras(config)
