+++ package_metadata.yaml
@@ -1,6 +1,6 @@
 __version__: "2.0.0"
-__name__: "lidar_multiclass"
-__url__: "https://github.com/IGNF/lidar-deep-segmentation"
+__name__: "myria3d"
+__url__: "https://github.com/IGNF/myria3d"
 __description__: "Multiclass Semantic Segmentation for Lidar Point Cloud"
 __author__: "Charles GAYDON"
 __copyright__: "2022, Institut National de l'Information Géographique et Forestière"
