+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-project_name = "Lidar-Deep-Segmentation"
+project_name = "Myria3D"
 author = "Charles GAYDON"
 contact = "charles.gaydon@gmail.com"
 license_file = LICENSE
@@ -27,7 +27,7 @@ python_files = tests/*
 log_cli = True
 addopts =
     # Always use coverage when running pytest
-    --cov "./lidar_multiclass/"
+    --cov "./myria3d/"
     # Percentage under which coverage is not reached 
     --cov-fail-under 70
     --cov-report html
