+++ setup_env/setup_env.sh
@@ -1,6 +1,3 @@
-set -e
-
-# SETUP
 # WARNING: you may need to adapt cudatoolkit version to your needs.
 # Or remove it altogether in case you do not need CUDA support at all
 
@@ -9,4 +6,4 @@ conda install -y mamba -n base -c conda-forge
 # Build it
 mamba env create -f setup_env/requirements.yml
 # activate it
-conda activate lidar_multiclass
+conda activate myria3d
