+++ tests/myria3d/data/test_transforms.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 import torch_geometric
-from lidar_multiclass.data.transforms import TargetTransform
+from myria3d.data.transforms import TargetTransform
 
 
 def test_TargetTransform_with_valid_config():
